requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
absl-py==0.13.0
alembic==1.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
asgiref==3.4.0; python_version >= '3.6'
astunparse==1.6.3
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
boto3==1.17.103
botocore==1.20.103; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
cached-property==1.5.2; python_version < '3.8'
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
clang==5.0
click==8.0.1; python_version >= '3.6'
cloudpickle==1.6.0; python_version >= '3.5'
databricks-cli==0.14.3
dataclasses==0.6
dill==0.3.4; python_version >= '2.7' and python_version != '3.0'
docker==5.0.0; python_version >= '3.6'
entrypoints==0.3; python_version >= '2.7'
evidently==0.1.19.dev0
fastapi==0.65.2
flask==2.0.1
flatbuffers==1.12
gast==0.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gitdb==4.0.7; python_version >= '3.4'
gitpython==3.1.18; python_version >= '3.6'
google-auth-oauthlib==0.4.4; python_version >= '3.6'
google-auth==1.32.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-pasta==0.2.0
graphviz==0.16
greenlet==1.1.0; python_version >= '3'
grpcio==1.38.1
gunicorn==20.1.0
h11==0.12.0; python_version >= '3.6'
h5py==3.1.0; python_version >= '3.6'
httptools==0.2.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==4.6.0; python_version < '3.8'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
joblib==1.0.1; python_version >= '3.6'
keras-preprocessing==1.1.2
keras==2.6.0rc0
llvmlite==0.36.0; python_version < '3.10' and python_version >= '3.6'
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markdown==3.3.4; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
mlflow==1.18.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
numba==0.53.1; python_version < '3.10' and python_version >= '3.6'
numpy==1.19.5; python_version == '3.7'
oauthlib==3.1.1; python_version >= '3.6'
opt-einsum==3.3.0; python_version >= '3.5'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas==1.3.0rc1
pandera==0.6.4
pathos==0.2.8
patsy==0.5.1
plotly==5.1.0; python_version >= '3.6'
pox==0.3.0
ppft==1.6.6.4
prometheus-client==0.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prometheus-flask-exporter==0.18.2
protobuf3-to-dict==0.1.5
protobuf==3.17.3
psutil==5.8.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
psycopg2==2.9.1
pyaml==20.4.0
pyarrow==4.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.8.2; python_full_version >= '3.6.1'
pydot==1.4.2
pyparsing==3.0.0b2; python_version >= '3.5'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
pyyaml==5.4.1
querystring-parser==1.2.4
requests-oauthlib==1.3.0
requests==2.25.1
retrying==1.3.3
rsa==4.7.2; python_version >= '3.6'
s3transfer==0.4.2
sagemaker-inference==1.5.5
sagemaker==2.47.2
scikit-learn==0.24.2
scikit-optimize==0.8.1
scipy==1.7.0; python_version < '3.10' and python_version >= '3.7'
shap==0.39.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
slicer==0.0.7; python_version >= '3.6'
smdebug-rulesconfig==1.0.1; python_version >= '2.7'
smmap==4.0.0; python_version >= '3.5'
sqlalchemy==1.4.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlparse==0.4.1; python_version >= '3.5'
starlette==0.14.2; python_version >= '3.6'
statsmodels==0.12.2; python_version >= '3.6'
tabulate==0.8.9
tenacity==7.0.0
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.0
tensorboard==2.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorflow==2.6.0rc0
termcolor==1.1.0
tf-estimator-nightly==2.6.0.dev2021062501
threadpoolctl==2.1.0; python_version >= '3.5'
tqdm==4.61.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.7.4.3; python_version < '3.8'
typing-inspect==0.7.1
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn[standard]==0.14.0
uvloop==0.15.2
watchgod==0.7
websocket-client==1.1.0; python_version >= '3.6'
websockets==9.1
werkzeug==2.0.1; python_version >= '3.6'
wheel==0.36.2; python_version >= '3'
wrapt==1.12.1
zipp==3.4.1; python_version >= '3.6'
